Rename --legacy argument to --archive-mode
Legacy (now archive) mode saves additional metadata and because
it uses a flat layout it prevents multiple downloads of the same
files in cases like this:
A
\-B
  \-file1
  \-file2
where file1 and file2 will get downloaded twice if a link to folder
A and folder B are both present in the download list. It shouldn't
be a problem for the average user but in a case when someone wants
to download a huge list of automatically scraped links it might
cause unnecessary wasted space.